⚡️ Modify : 학생 권한 탭 추가
- 학생 권한 탭 추가
