+++ src/pages/common/MyPage.tsx
@@ -11,6 +11,7 @@ const MyPage = () => {
 			isUseHeader: true,
 			isUseMenu: false,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
