+++ src/pages/common/SignUpPage.tsx
@@ -11,6 +11,7 @@ const SignUpPage = () => {
 			isUseHeader: true,
 			isUseMenu: false,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
