+++ src/pages/common/SplashPage.tsx
@@ -11,6 +11,7 @@ const SplashPage = () => {
 			isUseHeader: false,
 			isUseMenu: false,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
