+++ src/pages/educator/item/DetailItemPage.tsx
@@ -11,6 +11,7 @@ const DetailItemPage = () => {
 			isUseHeader: false,
 			isUseMenu: true,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
