+++ src/pages/educator/item/EditItemPage.tsx
@@ -11,6 +11,7 @@ const EditItemPage = () => {
 			isUseHeader: false,
 			isUseMenu: true,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
