+++ src/pages/educator/item/ItemPage.tsx
@@ -11,6 +11,7 @@ const ItemPage = () => {
 			isUseHeader: true,
 			isUseMenu: true,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
