+++ src/pages/educator/item/RegisterItemPage.tsx
@@ -11,6 +11,7 @@ const RegisterItemPage = () => {
 			isUseHeader: false,
 			isUseMenu: true,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
