+++ src/pages/educator/manageLearner/ManageLearnerPage.tsx
@@ -11,6 +11,7 @@ const ManageLearnerPage = () => {
 			isUseHeader: true,
 			isUseMenu: true,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
