+++ src/pages/educator/trade/DetailStockPage.tsx
@@ -11,6 +11,7 @@ const DetailTradeStockPage = () => {
 			isUseHeader: false,
 			isUseMenu: true,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
