+++ src/pages/educator/trade/EditStockPage.tsx
@@ -11,6 +11,7 @@ const EditStockPage = () => {
 			isUseHeader: false,
 			isUseMenu: true,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
