+++ src/pages/educator/trade/RegisterStockPage.tsx
@@ -11,6 +11,7 @@ const RegisterStockPage = () => {
 			isUseHeader: false,
 			isUseMenu: true,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
