+++ src/pages/post/DetailPostPage.tsx
@@ -11,6 +11,7 @@ const DetailPostPage = () => {
 			isUseHeader: false,
 			isUseMenu: true,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
