+++ src/pages/post/EditPostPage.tsx
@@ -11,6 +11,7 @@ const EditPostPage = () => {
 			isUseHeader: false,
 			isUseMenu: true,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
