+++ src/pages/student/stock/LearnerStockDetailPage.tsx
@@ -16,6 +16,7 @@ const LearnerStockDetailPage = () => {
 			isUseHeader: false,
 			isUseMenu: true,
 			isUseTab: false,
+			isUseItemTab: false,
 		});
 	}, []);
