+++ src/states/headerMenuUseState.ts
@@ -4,9 +4,15 @@ interface IHeaderMenuUseState {
 	readonly isUseHeader: boolean;
 	readonly isUseMenu: boolean;
 	readonly isUseTab: boolean;
+	readonly isUseItemTab: boolean;
 }
 
 export const headerMenuUseState = atom<IHeaderMenuUseState>({
-	default: { isUseHeader: false, isUseMenu: false, isUseTab: false },
+	default: {
+		isUseHeader: false,
+		isUseMenu: false,
+		isUseTab: false,
+		isUseItemTab: false,
+	},
 	key: "headerMenuUseState",
 });
